networkSecurity/components/dataValidation.py
from networkSecurity.entity.artifactEntity import DataIngestionArtifact, DataValidationArtifact
from networkSecurity.entity.configEntity import DataValidationConfig

from networkSecurity.exceptionHandling.exception import NetworkSecurityException
from networkSecurity.logging.logger import logging
from networkSecurity.constant.trainingPipeline import SCHEMA_FILE_PATH
from networkSecurity.utils.mainUtils.utils import read_yaml_file, write_yaml_file

from scipy.stats import ks_2samp
import pandas as pd
import os
import sys

class DataValidation:
    def __init__(self, data_validation_config: DataValidationConfig,
                 data_ingestion_artifact: DataIngestionArtifact):
        try:
            logging.info(f"{'>>'*20} Data Validation {'<<'*20}")
            self.data_validation_config = data_validation_config
            self.data_ingestion_artifact = data_ingestion_artifact
            self.schema_file_path = read_yaml_file(SCHEMA_FILE_PATH)
        except Exception as e:
            raise NetworkSecurityException(e, sys) from e
    
    @staticmethod
    def read_data(file_path) -> pd.DataFrame:
        try:
            return pd.read_csv(file_path)
        except Exception as e:
            raise NetworkSecurityException(e, sys) from e

    def validate_number_of_columns(self, dataframe: pd.DataFrame) -> bool:
        try:
            number_of_columns = len(self.schema_file_path["columns"])
            logging.info(f"Required number of columns: {number_of_columns}")
            logging.info(f"Dataframe has columns: {len(dataframe.columns)}")
            if len(dataframe.columns) == number_of_columns:
                return True
            return False
        except Exception as e:
            raise NetworkSecurityException(e, sys) from e

    def detect_data_drift(self, base_dataframe: pd.DataFrame,
                        current_dataframe: pd.DataFrame,
                        threshold: float = 0.05) -> bool:
        try:
            status=True
            drift_report = {}
            for column in base_dataframe.select_dtypes(include=["number"]).columns:
                base_data = base_dataframe[column].dropna()
                current_data = current_dataframe[column].dropna()
                is_same_dist = ks_2samp(base_data, current_data)
                if threshold <= is_same_dist.pvalue:
                    is_found = False
                else:
                    is_found = True
                    status=False
                drift_report.update({column:{"p_value":float(is_same_dist.pvalue),
                                             "drift_status":is_found}})
            drift_report_file_path = self.data_validation_config.drift_report_file_path
            dir_path = os.path.dirname(drift_report_file_path)
            os.makedirs(dir_path, exist_ok=True)
            write_yaml_file(file_path=drift_report_file_path, content=drift_report)
            return status

        except Exception as e:
            raise NetworkSecurityException(e, sys) from e

    def initiate_data_validation(self) -> DataValidationArtifact:
        try:
            train_file_path = self.data_ingestion_artifact.train_file_path
            test_file_path = self.data_ingestion_artifact.test_file_path

            #reading data from files
            train_df = DataValidation.read_data(train_file_path)
            test_df = DataValidation.read_data(test_file_path)

            #validating number of columns
            logging.info("Validating number of columns in training data")
            status = self.validate_number_of_columns(dataframe=train_df)
            if not status:
                raise Exception("Training data does not have the required number of columns")
            
            status = self.validate_number_of_columns(dataframe=test_df)
            if not status:
                raise Exception("Test data does not have the required number of columns")
            
            status = self.detect_data_drift(base_dataframe=train_df,
                                            current_dataframe=test_df)
            dir_path = os.path.dirname(self.data_validation_config.valid_train_file_path)
            os.makedirs(dir_path, exist_ok=True)

            train_df.to_csv(self.data_validation_config.valid_train_file_path, index=False)
            test_df.to_csv(self.data_validation_config.valid_test_file_path, index=False)

            data_validation_artifact = DataValidationArtifact(
                validation_status=status,
                valid_train_file_path=self.data_validation_config.valid_train_file_path,
                valid_test_file_path=self.data_validation_config.valid_test_file_path,
                invalid_train_file_path=None,
                invalid_test_file_path=None,
                drift_report_file_path=self.data_validation_config.drift_report_file_path
            )
            return data_validation_artifact 

        except Exception as e:
            raise NetworkSecurityException(e, sys) from e