requirements.txt
numpy
pandas
python-dotenv
setuptools
certifi
pymongo[srv]
pymongo
scikit-learn
dill
pyaml
##-e .